# recommender/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Summoner Recommender",
    "description": "API for recommending League of Legends summoners and managing recommendations.",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/matches/{summoner_name}": {
      "get": {
        "summary": "Get matches for a summoner",
        "description": "Returns all matches for the given summoner name.",
        "parameters": [
          {
            "name": "summoner_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The summoner's name to retrieve matches for"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of matches",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Match"
              }
            }
          },
          "404": {
            "description": "Summoner not found"
          }
        }
      }
    },
    "/champions": {
      "get": {
        "summary": "Get all champions",
        "description": "Returns a list of all League of Legends champions.",
        "responses": {
          "200": {
            "description": "A list of champions",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Champion"
              }
            }
          }
        }
      }
    },
    "/champion-rotations": {
      "get": {
        "summary": "Get free champion rotations",
        "description": "Returns the current free champion rotations.",
        "responses": {
          "200": {
            "description": "A dictionary of free champions, free champions for new players and max new player level.",
            "schema": {
              "type": "object",
              "properties": {
                "free_champions": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Champion"
                  }
                },
                "free_champions_for_new_players": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Champion"
                  }
                },
                "max_new_player_level": {
                  "type": "integer"
                }
              }
            }
          },
          "404": {
            "description": "Champion rotations not found"
          }
        }
      }
    },
    "/recommendations/{summoner_name}": {
      "get": {
        "summary": "Get recommendations for a summoner",
        "description": "Returns a list of recommended summoners based on the given summoner name.",
        "parameters": [
          {
            "name": "summoner_name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "number_of_recommendations",
            "in": "query",
            "required": false,
            "type": "integer",
            "default": 10,
            "description": "The number of recommendations to return"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of summoner recommendations",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Recommendation"
              }
            }
          },
          "404": {
            "description": "Summoner not found"
          }
        }
      }
    },
    "/recommendation/update/{summoner_name}": {
      "post": {
        "summary": "Update recommendation status",
        "description": "Updates the status of a recommendation for a summoner.",
        "parameters": [
          {
            "name": "summoner_name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RecommendationUpdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recommendation status updated successfully",
            "schema": {
              "$ref": "#/definitions/UpdateResponse"
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Summoner not found"
          }
        }
      }
    },
    "/profile": {
      "post": {
        "summary": "Create a new profile and add to the database",
        "description": "Creates a new profile for a summoner.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile created successfully"
          },
          "400": {
            "description": "Invalid request"
          },
          "409": {
            "description": "Profile already exists"
          }
        }
      }
    }
  },
  "definitions": {
    "Match": {
      "type": "object",
      "properties": {
        "summoner1_id": {
          "type": "integer"
        },
        "summoner2_id": {
          "type": "integer"
        },
        "match_date": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "Champion": {
      "type": "object",
      "properties": {
        "champion_id": {
          "type": "integer"
        },
        "champion_name": {
          "type": "string"
        }
      }
    },
    "Recommendation": {
      "type": "object",
      "properties": {
        "age": {
          "type": "integer"
        },
        "country": {
          "type": "string"
        },
        "favorite_champion_id": {
          "type": "integer"
        },
        "favourite_line": {
          "type": "string"
        },
        "languages": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "level": {
          "type": "integer"
        },
        "long_description": {
          "type": "string"
        },
        "losses": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "preferred_champion_ids_and_lines": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "champion_id": {
                "type": "integer"
              },
              "line": {
                "type": "string"
              }
            }
          }
        },
        "short_description": {
          "type": "string"
        },
        "tier": {
          "type": "string"
        },
        "win_rate": {
          "type": "number"
        },
        "wins": {
          "type": "integer"
        }
      }
    },
    "Profile": {
      "type": "object",
      "required": [
        "summoner_name",
        "sex",
        "country",
        "languages",
        "age",
        "preferred_champion_ids_and_lines",
        "favourite_champion_id",
        "favourite_line",
        "long_description",
        "short_description",
        "icon_id"
      ],
      "properties": {
        "summoner_name": {
          "type": "string"
        },
        "sex": {
          "type": "char"
        },
        "country": {
          "type": "string"
        },
        "languages": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "age": {
          "type": "integer"
        },
        "preferred_champion_ids_and_lines": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "champion_id": {
                "type": "integer"
              },
              "line": {
                "type": "string"
              }
            }
          }
        },
        "favourite_champion_id": {
          "type": "integer"
        },
        "favourite_line": {
          "type": "string"
        },
        "long_description": {
          "type": "string"
        },
        "short_description": {
          "type": "string"
        },
        "icon_id": {
          "type": "integer"
        }
      }
    },
    "RecommendationUpdate": {
      "type": "object",
      "required": [
        "recommendation",
        "status"
      ],
      "properties": {
        "recommendation": {
          "type": "string",
          "description": "The recommended summoner's name"
        },
        "status": {
          "type": "string",
          "enum": [
            "accept",
            "reject"
          ],
          "description": "The status of the recommendation"
        }
      }
    },
    "UpdateResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "match": {
          "type": "boolean"
        }
      }
    }
  }
}